+++ R/downloadStudy.R
@@ -69,10 +69,19 @@
     .manageLocalFile(cancer_study_id, tmpFile)
 }
 
+# previously http://download.cbioportal.org
+.url_location <- "https://cbioportal-datahub.s3.amazonaws.com"
+
 #' Download and cache study dataset
 #'
 #' Provide a `cancer_study_id` from the `studiesTable` and retrieve
-#' the study tarball from cBioPortal
+#' the study tarball from cBioPortal. This function is used by `cBioDataPack`
+#' to retrieve the tarball datasets and cache them. As stated in
+#' `?cBioDataPack` not all studies are currently working as
+#' `MultiAssayExperiment` representations. As of July 2020, about ~80% of
+#' datasets can be successfully imported into the `MultiAssayExperiment` data
+#' class. Please open an issue if you would like the team to prioritize a
+#' study.
 #'
 #' @param cancer_study_id The cBioPortal study identifier as in
 #'     \url{https://cbioportal.org/webAPI}
@@ -84,14 +93,24 @@
 #' @param force logical (default FALSE) whether to force re-download data from
 #' remote location
 #'
+#' @param url_location (default "https://cbioportal-datahub.s3.amazonaws.com")
+#' the URL location for downloading packaged data. Can be set using the
+#' 'cBio_URL' option
+#'
 #' @return The file location of the data tarball
 #'
-#' @keywords internal
-downloadStudy <- function(cancer_study_id, use_cache = TRUE, force = FALSE)
+#' @md
+#'
+#' @examples
+#'
+#' downloadStudy("acc_tcga")
+#'
+#' @export
+downloadStudy <- function(cancer_study_id, use_cache = TRUE, force = FALSE,
+    url_location = getOption("cBio_URL", .url_location))
 {
     .validStudyID(cancer_study_id)
 
-    url_location <- "http://download.cbioportal.org"
     url_file <- file.path(url_location, paste0(cancer_study_id, ".tar.gz"))
 
     if (is.character(use_cache) && length(use_cache) == 1L)